requirements.txt
aiohttp==3.8.4
asyncio==3.4.3
python-dateutil==2.8.2